First pass of cleanup and changes to match style
Major of this was just to match my personal take on mod development. Though I do feel like there was an over use of classes. A lot of classes were very unneeded in the way they were implemented. Such as the ModBlocks.class. Which could have at minimal included the item and block registry event.

The only issues noticed was depending on the client to verify teleportation. This was very easy to exploit client side to teleport anywhere in world. As the server never validated the to and from location. The solution for this was to use the packet to only trigger the server. Then let the server do the actual work.

The other issue was using the display name method to show the block name. This could have been done with ItemWool. Which would append the translation key to the end. Not a big issue but needed to be resolved in order to allow translations to work correctly. As English/German allows 'Red Block' but other languages do 'block or red'. It gets even more complex for non-western languages.